refactor(appauthconfig): move params struct and app session id generation
